fix cpp format, change proof system folder structure
