--- zokrates_core/src/proof_system/bn128/utils/mod.rs
@@ -1,6 +1,0 @@
-pub mod bellman;
-#[cfg(feature = "libsnark")]
-pub mod ffi;
-#[cfg(feature = "libsnark")]
-pub mod libsnark;
-pub mod solidity;
