+++ zokrates_core/src/proof_system/libsnark/ffi.rs
