+++ zokrates_core/src/proof_system/libsnark/mod.rs
@@ -1,3 +1,7 @@
+mod ffi;
+pub mod gm17;
+pub mod pghr13;
+
 use flat_absy::FlatVariable;
 use ir::{self, Statement};
 use std::cmp::max;
