+++ zokrates_core/src/proof_system/mod.rs
@@ -1,10 +1,8 @@
-mod bn128;
+pub mod bellman;
+#[cfg(feature = "libsnark")]
+pub mod libsnark;
 
-pub use self::bn128::G16;
-#[cfg(feature = "libsnark")]
-pub use self::bn128::GM17;
-#[cfg(feature = "libsnark")]
-pub use self::bn128::PGHR13;
+mod solidity;
 
 use crate::ir;
 use serde::de::DeserializeOwned;
@@ -54,6 +52,24 @@ impl<T: Serialize + DeserializeOwned> Proof<T> {
             inputs,
             raw,
         }
+    }
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct G1Affine(String, String);
+
+#[derive(Serialize, Deserialize)]
+pub struct G2Affine(G1Affine, G1Affine);
+
+impl ToString for G1Affine {
+    fn to_string(&self) -> String {
+        format!("{}, {}", self.0, self.1)
+    }
+}
+
+impl ToString for G2Affine {
+    fn to_string(&self) -> String {
+        format!("[{}], [{}]", self.0.to_string(), self.1.to_string())
     }
 }
