+++ zokrates_core/src/proof_system/solidity.rs
